cambios en logica de archivo para pasar a utf-8 y se crea api para obtener derecho de cuenta corriente estructurado cabecera detalle agrupado por causante

+++ solicitudes-pago/src/main/java/cl/gob/ips/solicitudes_pago/dto/DerechoCausanteDTO.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class DerechoCausanteDTO {
     private int id;
-    private long rutCausante;
+    private int rutCausante;
     private String dvCausante;
-    private long rutBeneficiario;
+    private int rutBeneficiario;
     private String dvBeneficiario;
     private int periodo;
     private int tipoMovimientoId;
